refactor: cleanup `ProposalsDetailsMenu`
